+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 ## X.Y.Z - changes pending release 
 
+### PaymentSheet
+* [Fixed] Fixed build for visionOS 26.2 SDK.
+
 ## 25.6.1 2026-02-03
 
 ### Identity